Foreign.CUDA.Analysis.Device

Description

Common device functions

Documentation

data Compute

GPU compute capability, major and minor revision number respectively.

Constructors

Compute !Int !Int

data ComputeMode

The compute mode the device is currently in

Constructors

Default

Prohibited

ExclusiveProcess

data DeviceProperties

The properties of a compute device

Constructors

DeviceProperties

Fields

deviceName :: !String

Identifier

computeCapability :: !Compute

Supported compute capability

totalGlobalMem :: !Int64

Available global memory on the device in bytes

totalConstMem :: !Int64

Available constant memory on the device in bytes

sharedMemPerBlock :: !Int64

Available shared memory per block in bytes

regsPerBlock :: !Int

32-bit registers per block

warpSize :: !Int

Warp size in threads (SIMD width)

maxThreadsPerBlock :: !Int

Maximum number of threads per block

maxThreadsPerMultiProcessor :: !Int

Maximum number of threads per multiprocessor

maxBlockSize :: !(Int, Int, Int)

Maximum size of each dimension of a block

maxGridSize :: !(Int, Int, Int)

Maximum size of each dimension of a grid

maxTextureDim1D :: !Int

Maximum texture dimensions

maxTextureDim2D :: !(Int, Int)

maxTextureDim3D :: !(Int, Int, Int)

clockRate :: !Int

Clock frequency in kilohertz

multiProcessorCount :: !Int

Number of multiprocessors on the device

memPitch :: !Int64

Maximum pitch in bytes allowed by memory copies

memBusWidth :: !Int

Global memory bus width in bits

memClockRate :: !Int

Peak memory clock frequency in kilohertz

textureAlignment :: !Int64

Alignment requirement for textures

computeMode :: !ComputeMode

deviceOverlap :: !Bool

Device can concurrently copy memory and execute a kernel

concurrentKernels :: !Bool

Device can possibly execute multiple kernels concurrently

eccEnabled :: !Bool

Device supports and has enabled error correction

asyncEngineCount :: !Int

Number of asynchronous engines

cacheMemL2 :: !Int

Size of the L2 cache in bytes

pciInfo :: !PCI

PCI device information for the device

tccDriverEnabled :: !Bool

Whether this is a Tesla device using the TCC driver

kernelExecTimeoutEnabled :: !Bool

Whether there is a runtime limit on kernels

integrated :: !Bool

As opposed to discrete

canMapHostMemory :: !Bool

Device can use pinned memory

unifiedAddressing :: !Bool

Device shares a unified address space with the host

streamPriorities :: !Bool

Device supports stream priorities

globalL1Cache :: !Bool

Device supports caching globals in L1 cache

localL1Cache :: !Bool

Device supports caching locals in L1 cache

managedMemory :: !Bool

Device supports allocating managed memory on this system

multiGPUBoard :: !Bool

Device is on a multi-GPU board

multiGPUBoardGroupID :: !Int

Unique identifier for a group of devices associated with the same board

preemption :: !Bool

Device supports compute pre-emption

singleToDoublePerfRatio :: !Int

Ratio of single precision performance (in floating-point operations per second) to double precision performance

cooperativeLaunch :: !Bool

Device supports launching cooperative kernels

cooperativeLaunchMultiDevice :: !Bool

Device can participate in cooperative multi-device kernels

data DeviceResources

Constructors

DeviceResources

Fields

threadsPerWarp :: !Int

Warp size

coresPerMP :: !Int

Number of SIMD arithmetic units per multiprocessor

warpsPerMP :: !Int

Maximum number of in-flight warps per multiprocessor

threadsPerMP :: !Int

Maximum number of in-flight threads on a multiprocessor

threadBlocksPerMP :: !Int

Maximum number of thread blocks resident on a multiprocessor

sharedMemPerMP :: !Int

Total amount of shared memory per multiprocessor (bytes)

maxSharedMemPerBlock :: !Int

Maximum amount of shared memory per thread block (bytes)

regFileSizePerMP :: !Int

Total number of registers in a multiprocessor

maxRegPerBlock :: !Int

Maximum number of registers per block

regAllocUnit :: !Int

Register allocation unit size

regAllocationStyle :: !Allocation

How multiprocessor resources are divided (register allocation granularity)

maxRegPerThread :: !Int

Maximum number of registers per thread

sharedMemAllocUnit :: !Int

Shared memory allocation unit size (bytes)

warpAllocUnit :: !Int

Warp allocation granularity

warpRegAllocUnit :: !Int

Warp register allocation granularity

maxGridsPerDevice :: !Int

Maximum number of resident grids per device (concurrent kernels)

data Allocation

Constructors

Warp

Block

data PCI

Constructors

PCI

Fields

busID :: !Int

PCI bus ID of the device

deviceID :: !Int

PCI device ID

domainID :: !Int

PCI domain ID

deviceResources :: DeviceProperties -> DeviceResources

Extract some additional hardware resource limitations for a given device.

describe :: Describe a => a -> String