Foreign.CUDA.Driver.Context

Description

Context management for low-level driver interface

Context Management

newtype Context

A device context

Constructors

Context

Fields

useContext :: Ptr ()

data ContextFlag

Context creation flags

Constructors

SchedAuto

SchedSpin

SchedYield

SchedBlockingSync

BlockingSync

Deprecated: use SchedBlockingSync instead

SchedMask

MapHost

LmemResizeToMax

FlagsMask

create :: Device -> [ContextFlag] -> IO Context

Create a new CUDA context and associate it with the calling thread

attach :: Context -> [ContextFlag] -> IO ()

Deprecated: deprecated as of CUDA-4.0

Increments the usage count of the context. API: no context flags are currently supported, so this parameter must be empty.

detach :: Context -> IO ()

Deprecated: deprecated as of CUDA-4.0

Detach the context, and destroy if no longer used

destroy :: Context -> IO ()

Destroy the specified context. This fails if the context is more than a single attachment (including that from initial creation).

device :: IO Device

Return the device of the currently active context

pop :: IO Context

Pop the current CUDA context from the CPU thread. The context must have a single usage count (matching calls to attach and detach). If successful, the new context is returned, and the old may be attached to a different CPU.

push :: Context -> IO ()

Push the given context onto the CPU's thread stack of current contexts. The context must be floating (via pop), i.e. not attached to any thread.

sync :: IO ()

Block until the device has completed all preceding requests

get :: IO Context

Return the context bound to the calling CPU thread. Requires cuda-4.0.

set :: Context -> IO ()

Bind the specified context to the calling thread. Requires cuda-4.0.

Peer Access

data PeerFlag

Possible option values for direct peer memory access

accessible :: Device -> Device -> IO Bool

Queries if the first device can directly access the memory of the second. If direct access is possible, it can then be enabled with add. Requires cuda-4.0.

add :: Context -> [PeerFlag] -> IO ()

If the devices of both the current and supplied contexts support unified addressing, then enable allocations in the supplied context to be accessible by the current context. Requires cuda-4.0.

remove :: Context -> IO ()

Disable direct memory access from the current context to the supplied context. Requires cuda-4.0.

Cache Configuration

data Cache

Device cache configuration preference

Constructors

PreferNone

PreferShared

PreferL1

PreferEqual

data Limit

Device limits flags

Constructors

StackSize

PrintfFifoSize

MallocHeapSize

DevRuntimeSyncDepth

DevRuntimePendingLaunchCount

Max

getLimit :: Limit -> IO Int

Query compute 2.0 call stack limits. Requires cuda-3.1.

setLimit :: Limit -> Int -> IO ()

Specify the size of the call stack, for compute 2.0 devices. Requires cuda-3.1.

setCacheConfig :: Cache -> IO ()

On devices where the L1 cache and shared memory use the same hardware resources, this sets the preferred cache configuration for the current context. This is only a preference. Requires cuda-3.2.