Foreign.CUDA.Driver.Context.Peer

Description

Direct peer context access functions for the low-level driver interface.

Since: CUDA-4.0

Peer Access

data PeerFlag

Possible option values for direct peer memory access

accessible :: Device -> Device -> IO Bool

Queries if the first device can directly access the memory of the second. If direct access is possible, it can then be enabled with add.

Requires CUDA-4.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PEER__ACCESS.html#group__CUDA__PEER__ACCESS_1g496bdaae1f632ebfb695b99d2c40f19e

add :: Context -> [PeerFlag] -> IO ()

If the devices of both the current and supplied contexts support unified addressing, then enable allocations in the supplied context to be accessible by the current context.

Note that access is unidirectional, and in order to access memory in the current context from the peer context, a separate symmetric call to add is required.

Requires CUDA-4.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PEER__ACCESS.html#group__CUDA__PEER__ACCESS_1g0889ec6728e61c05ed359551d67b3f5a

remove :: Context -> IO ()

Disable direct memory access from the current context to the supplied peer context, and unregisters any registered allocations.

Requires CUDA-4.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PEER__ACCESS.html#group__CUDA__PEER__ACCESS_1g5b4b6936ea868d4954ce4d841a3b4810