Foreign.CUDA.Driver.Device

Description

Device management for low-level driver interface

Device Management

newtype Device

A CUDA device

Constructors

Device

Fields

useDevice :: CInt

data DeviceProperties

The properties of a compute device

Constructors

DeviceProperties

Fields

deviceName :: !String

Identifier

computeCapability :: !Compute

Supported compute capability

totalGlobalMem :: !Int64

Available global memory on the device in bytes

totalConstMem :: !Int64

Available constant memory on the device in bytes

sharedMemPerBlock :: !Int64

Available shared memory per block in bytes

regsPerBlock :: !Int

32-bit registers per block

warpSize :: !Int

Warp size in threads (SIMD width)

maxThreadsPerBlock :: !Int

Maximum number of threads per block

maxThreadsPerMultiProcessor :: !Int

Maximum number of threads per multiprocessor

maxBlockSize :: !(Int, Int, Int)

Maximum size of each dimension of a block

maxGridSize :: !(Int, Int, Int)

Maximum size of each dimension of a grid

maxTextureDim1D :: !Int

Maximum texture dimensions

maxTextureDim2D :: !(Int, Int)

maxTextureDim3D :: !(Int, Int, Int)

clockRate :: !Int

Clock frequency in kilohertz

multiProcessorCount :: !Int

Number of multiprocessors on the device

memPitch :: !Int64

Maximum pitch in bytes allowed by memory copies

memBusWidth :: !Int

Global memory bus width in bits

memClockRate :: !Int

Peak memory clock frequency in kilohertz

textureAlignment :: !Int64

Alignment requirement for textures

computeMode :: !ComputeMode

deviceOverlap :: !Bool

Device can concurrently copy memory and execute a kernel

concurrentKernels :: !Bool

Device can possibly execute multiple kernels concurrently

eccEnabled :: !Bool

Device supports and has enabled error correction

asyncEngineCount :: !Int

Number of asynchronous engines

cacheMemL2 :: !Int

Size of the L2 cache in bytes

pciInfo :: !PCI

PCI device information for the device

tccDriverEnabled :: !Bool

Whether this is a Tesla device using the TCC driver

kernelExecTimeoutEnabled :: !Bool

Whether there is a runtime limit on kernels

integrated :: !Bool

As opposed to discrete

canMapHostMemory :: !Bool

Device can use pinned memory

unifiedAddressing :: !Bool

Device shares a unified address space with the host

data DeviceAttribute

Device attributes

Constructors

MaxThreadsPerBlock

MaxBlockDimX

MaxBlockDimY

MaxBlockDimZ

MaxGridDimX

MaxGridDimY

MaxGridDimZ

MaxSharedMemoryPerBlock

SharedMemoryPerBlock

TotalConstantMemory

WarpSize

MaxPitch

MaxRegistersPerBlock

RegistersPerBlock

ClockRate

TextureAlignment

GpuOverlap

MultiprocessorCount

KernelExecTimeout

Integrated

CanMapHostMemory

ComputeMode

MaximumTexture1dWidth

MaximumTexture2dWidth

MaximumTexture2dHeight

MaximumTexture3dWidth

MaximumTexture3dHeight

MaximumTexture3dDepth

MaximumTexture2dLayeredWidth

MaximumTexture2dArrayWidth

MaximumTexture2dLayeredHeight

MaximumTexture2dArrayHeight

MaximumTexture2dLayeredLayers

MaximumTexture2dArrayNumslices

SurfaceAlignment

ConcurrentKernels

EccEnabled

PciBusId

PciDeviceId

TccDriver

MemoryClockRate

GlobalMemoryBusWidth

L2CacheSize

MaxThreadsPerMultiprocessor

AsyncEngineCount

UnifiedAddressing

MaximumTexture1dLayeredWidth

MaximumTexture1dLayeredLayers

PciDomainId

data Compute

GPU compute capability, major and minor revision number respectively.

Constructors

Compute !Int !Int

data ComputeMode

The compute mode the device is currently in

Constructors

Default

Exclusive

Prohibited

ExclusiveProcess

data InitFlag

Possible option flags for CUDA initialisation. Dummy instance until the API exports actual option values.

initialise :: [InitFlag] -> IO ()

Initialise the CUDA driver API. This must be called before any other driver function.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__INITIALIZE.html#group__CUDA__INITIALIZE_1g0a2f1517e1bd8502c7194c3a8c134bc3

capability :: Device -> IO Compute

Return the compute compatibility revision supported by the device

device :: Int -> IO Device

Return a handle to the compute device at the given ordinal.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__DEVICE.html#group__CUDA__DEVICE_1g8bdd1cc7201304b01357b8034f6587cb

attribute :: Device -> DeviceAttribute -> IO Int

Return the selected attribute for the given device.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__DEVICE.html#group__CUDA__DEVICE_1g9c3e1414f0ad901d3278a4d6645fc266

count :: IO Int

Return the number of device with compute capability > 1.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__DEVICE.html#group__CUDA__DEVICE_1g52b5ce05cb8c5fb6831b2c0ff2887c74

name :: Device -> IO String

The identifying name of the device.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__DEVICE.html#group__CUDA__DEVICE_1gef75aa30df95446a845f2a7b9fffbb7f

props :: Device -> IO DeviceProperties

Return the properties of the selected device

totalMem :: Device -> IO Int64

The total memory available on the device (bytes).

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__DEVICE.html#group__CUDA__DEVICE_1gc6a0d6551335a3780f9f3c967a0fde5d