Foreign.CUDA.Driver.Event

Description

Event management for low-level driver interface

Event Management

newtype Event

Events are markers that can be inserted into the CUDA execution stream and later queried.

Constructors

Event

Fields

useEvent :: Ptr ()

data EventFlag

Event creation flags

Constructors

Default

BlockingSync

DisableTiming

Interprocess

data WaitFlag

create :: [EventFlag] -> IO Event

Create a new event

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g450687e75f3ff992fe01662a43d9d3db

destroy :: Event -> IO ()

Destroy an event

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g593ec73a8ec5a5fc031311d3e4dca1ef

elapsedTime :: Event -> Event -> IO Float

Determine the elapsed time (in milliseconds) between two events

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1gdfb1178807353bbcaa9e245da497cf97

query :: Event -> IO Bool

Determines if a event has actually been recorded

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g6f0704d755066b0ee705749ae911deef

record :: Event -> Maybe Stream -> IO ()

Record an event once all operations in the current context (or optionally specified stream) have completed. This operation is asynchronous.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g95424d3be52c4eb95d83861b70fb89d1

wait :: Event -> Maybe Stream -> [WaitFlag] -> IO ()

Makes all future work submitted to the (optional) stream wait until the given event reports completion before beginning execution. Synchronisation is performed on the device, including when the event and stream are from different device contexts.

Requires CUDA-3.2.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g6a898b652dfc6aa1d5c8d97062618b2f

block :: Event -> IO ()

Wait until the event has been recorded

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g9e520d34e51af7f5375610bca4add99c