Foreign.CUDA.Driver.Graph.Build

Description

Graph construction functions for the low-level driver interface

Requires CUDA-10

Documentation

newtype Graph

Constructors

Graph

Fields

useGraph :: Ptr ()

newtype Node

Constructors

Node

Fields

useNode :: Ptr ()

data NodeType

Constructors

Kernel

Memcpy

Memset

Host

Subgraph

Empty

Count

type HostCallback = FunPtr (Ptr () -> IO ())

Callback function executed on the host

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TYPES.html#group__CUDA__TYPES_1g262cd3570ff5d396db4e3dabede3c355

Since: 0.10.0.0

create :: [GraphFlag] -> IO Graph

Create an empty task graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gd885f719186010727b75c3315f865fdf

Since: 0.10.0.0

destroy :: Graph -> IO ()

Destroy a graph, as well as all of its nodes

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g718cfd9681f078693d4be2426fd689c8

Since: 0.10.0.0

clone :: Graph -> IO Graph

Clone a graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g3603974654e463f2231c71d9b9d1517e

Since: 0.10.0.0

remove :: Node -> IO ()

Remove a node from the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g00ed16434d983d8f0011683eacaf19b9

Since: 0.10.0.0

Construction

addChild :: Graph -> Graph -> [Node] -> IO Node

Create a child graph node and add it to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g3f27c2e56e3d568b09f00d438e61ceb1

Since: 0.10.0.0

addEmpty :: Graph -> [Node] -> IO Node

Create an empty node and add it to the graph.

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g8a8681dbe97dbbb236ea5ebf3abe2ada

Since: 0.10.0.0

addHost :: Graph -> [Node] -> HostCallback -> Ptr () -> IO Node

Creates a host execution node and adds it to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g1ba15c2fe1afb8897091ecec4202b597

Since: 0.10.0.0

addKernel

Arguments

:: Graph

-> [Node]

-> Fun

-> (Int, Int, Int)
grid dimension

-> (Int, Int, Int)
thread block dimensions

-> Int
shared memory (bytes)

-> [FunParam]

-> IO Node

Create a kernel execution node and adds it to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g886a9096293238937f2f3bc7f2d57635

Since: 0.10.0.0

addMemcpy

Arguments

:: Graph

-> [Node]

-> Context

-> Int
srcXInBytes

-> Int
srcY

-> Int
srcZ

-> Int
srcLOD

-> MemoryType
source memory type

-> Ptr a
source ptr

-> Int
srcPitch

-> Int
srcHeight

-> Int
dstXInBytes

-> Int
dstY

-> Int
dstZ

-> Int
dstLOD

-> MemoryType
destination memory type

-> Ptr a
destination ptr

-> Int
dstPitch

-> Int
dstHeight

-> Int
widthInBytes

-> Int
height

-> Int
depth

-> IO Node

Create a memcpy node and add it to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gdd521e1437c1c3ea8822f66a32ff1f94

Since: 0.10.0.0

addMemset

Arguments

:: Storable a

=> Graph

-> [Node]

-> Context

-> DevicePtr a

-> a

-> Int
height

-> Int
pitch

-> Int
width

-> IO Node

Create a memset node and add it to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gac7f59961798f14a9f94f9f6b53cc3b7

Since: 0.10.0.0

addDependencies :: Graph -> [(Node, Node)] -> IO ()

Add dependency edges to the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g81bf1a6965f881be6ad8d21cfe0ee44f

Since: 0.10.0.0

removeDependencies :: Graph -> [(Node, Node)] -> IO ()

Remove dependency edges from the graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g8ab696a6b3ccd99db47feba7e97fb579

Since: 0.10.0.0

Querying

getType :: Node -> IO NodeType

Return the type of a node

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gdb1776d97aa1c9d5144774b29e4b8c3e

Since: 0.10.0.0

getChildGraph :: Node -> IO Graph

Retrieve the embedded graph of a child sub-graph node

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gbe9fc9267316b3778ef0db507917b4fd

Since: 0.10.0.0

getEdges :: Graph -> IO [(Node, Node)]

Return a graph's dependency edges

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g2b7bd71b0b2b8521f141996e0975a0d7

Since: 0.10.0.0

getNodes :: Graph -> IO [Node]

Return a graph's nodes

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gfa35a8e2d2fc32f48dbd67ba27cf27e5

Since: 0.10.0.0

getRootNodes :: Graph -> IO [Node]

Returns the root nodes of a graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gf8517646bd8b39ab6359f8e7f0edffbd

Since: 0.10.0.0

getDependencies :: Node -> IO [Node]

Return the dependencies of a node

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g048f4c0babcbba64a933fc277cd45083

Since: 0.10.0.0

getDependents :: Node -> IO [Node]

Return a node's dependent nodes

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g4b73d9e3b386a9c0b094a452b8431f59

Since: 0.10.0.0

findInClone :: Node -> Graph -> IO Node

Find a cloned version of a node

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gf21f6c968e346f028737c1118bfd41c2

Since: 0.10.0.0