Foreign.CUDA.Driver.Graph.Capture

Description

Requires CUDA-10

Documentation

data Status

Constructors

None

Active

Invalidated

data Mode

start :: Stream -> Mode -> IO ()

Begin graph capture on a stream

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1gea22d4496b1c8d02d0607bb05743532f

Since: 0.10.0.0

stop :: Stream -> IO Graph

End graph capture on a stream

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g03dab8b2ba76b00718955177a929970c

Since: 0.10.0.0

status :: Stream -> IO Status

Return a stream's capture status

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g37823c49206e3704ae23c7ad78560bca

Since: 0.10.0.0

info :: Stream -> IO (Status, Int64)

Query the capture status of a stream and get an id for the capture sequence, which is unique over the lifetime of the process.

Requires CUDA-10.1

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g13145ece1d79a1d79a1d22abb9663216

Since: 0.10.1.0

mode :: Mode -> IO Mode

Set the stream capture interaction mode for this thread. Return the previous value.

Requires CUDA-10.1

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g378135b262f02a43a7caeab239ae493d

Since: 0.10.1.0