Foreign.CUDA.Driver.Graph.Exec

Description

Graph execution functions for the low-level driver interface

Requires CUDA-10

Documentation

newtype Executable

Constructors

Executable

Fields

useExecutable :: Ptr ()

Execution

launch :: Executable -> Stream -> IO ()

Execute a graph in the given stream. Only one instance may execute at a time; to execute a graph concurrently, it must be instantiated into multiple executables.

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g6b2dceb3901e71a390d2bd8b0491e471

Since: 0.10.0.0

instantiate :: Graph -> IO Executable

Instantiate the task graph description of a program into an executable graph.

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1g433ae118a751c9f2087f53d7add7bc2c

Since: 0.10.0.0

destroy :: Executable -> IO ()

Destroy an executable graph

Requires CUDA-10.0

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1ga32ad4944cc5d408158207c978bc43a7

Since: 0.10.0.0

setKernel

Arguments

:: Executable

-> Node

-> Fun

-> (Int, Int, Int)
grid dimension

-> (Int, Int, Int)
thread block dimensions

-> Int
shared memory (bytes)

-> [FunParam]

-> IO ()

Update the parameters for a kernel node in the given executable graph

Requires CUDA-10.1

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__GRAPH.html#group__CUDA__GRAPH_1gd84243569e4c3d6356b9f2eea20ed48c

Since: 0.10.1.0