Foreign.CUDA.Driver.Module.Link

Description

Module linking for low-level driver interface

Since CUDA-5.5

JIT module linking

data LinkState

A pending JIT linker state

data JITOption

Just-in-time compilation and linking options

Constructors

MaxRegisters !Int
maximum number of registers per thread

ThreadsPerBlock !Int
number of threads per block to target for

OptimisationLevel !Int
level of optimisation to apply (1-4, default 4)

Target !Compute
compilation target, otherwise determined from context

FallbackStrategy !JITFallback
fallback strategy if matching cubin not found

GenerateDebugInfo
generate debug info (-g) (requires cuda >= 5.5)

GenerateLineInfo
generate line number information (-lineinfo) (requires cuda >= 5.5)

Verbose
verbose log messages (requires cuda >= 5.5)

data JITInputType

Device code formats that can be used for online linking

create :: [JITOption] -> IO LinkState

Create a pending JIT linker invocation. The returned LinkState should be destroyed once no longer needed. The device code machine size will match the calling application.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g86ca4052a2fab369cb943523908aa80d

destroy :: LinkState -> IO ()

Destroy the state of a JIT linker invocation.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g01b7ae2a34047b05716969af245ce2d9

complete :: LinkState -> IO Module

Complete a pending linker invocation and load the current module. The link state will be destroyed.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g818fcd84a4150a997c0bba76fef4e716

addFile :: LinkState -> FilePath -> JITInputType -> [JITOption] -> IO ()

Add an input file to a pending linker invocation.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g1224c0fd48d4a683f3ce19997f200a8c

addData :: LinkState -> ByteString -> JITInputType -> [JITOption] -> IO ()

Add an input to a pending linker invocation.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g3ebcd2ccb772ba9c120937a2d2831b77

addDataFromPtr :: LinkState -> Int -> Ptr Word8 -> JITInputType -> [JITOption] -> IO ()

As addData, but read the specified number of bytes of image data from the given pointer.