Foreign.CUDA.Driver.Module.Query

Description

Querying module attributes for low-level driver interface

Querying module inhabitants

getFun :: Module -> ShortByteString -> IO Fun

Returns a function handle.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1ga52be009b0d4045811b30c965e1cb2cf

getPtr :: Module -> ShortByteString -> IO (DevicePtr a, Int)

Return a global pointer, and size of the global (in bytes).

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1gf3e43672e26073b1081476dbf47a86ab

getTex :: Module -> ShortByteString -> IO Texture

Return a handle to a texture reference. This texture reference handle should not be destroyed, as the texture will be destroyed automatically when the module is unloaded.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MODULE.html#group__CUDA__MODULE_1g9607dcbf911c16420d5264273f2b5608