Foreign.CUDA.Driver.Profiler

Description

Profiler control for low-level driver interface

Documentation

data OutputMode

Profiler output mode

Constructors

KeyValuePair

CSV

initialise

Arguments

:: FilePath
configuration file that itemises which counters and/or options to profile

-> FilePath
output file where profiling results will be stored

-> OutputMode

-> IO ()

Initialise the CUDA profiler.

The configuration file is used to specify profiling options and profiling counters. Refer to the "Compute Command Line Profiler User Guide" for supported profiler options and counters.

Note that the CUDA profiler can not be initialised with this function if another profiling tool is already active.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PROFILER.html#group__CUDA__PROFILER

start :: IO ()

Begin profiling collection by the active profiling tool for the current context. If profiling is already enabled, then this has no effect.

start and stop can be used to programatically control profiling granularity, by allowing profiling to be done only on selected pieces of code.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PROFILER.html#group__CUDA__PROFILER_1g8a5314de2292c2efac83ac7fcfa9190e

stop :: IO ()

Stop profiling collection by the active profiling tool for the current context, and force all pending profiler events to be written to the output file. If profiling is already inactive, this has no effect.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__PROFILER.html#group__CUDA__PROFILER_1g4d8edef6174fd90165e6ac838f320a5f