Foreign.CUDA.Driver.Stream

Description

Stream management for low-level driver interface

Stream Management

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

type StreamPriority = Int

Priority of an execution stream. Work submitted to a higher priority stream may preempt execution of work already executing in a lower priority stream. Lower numbers represent higher priorities.

type StreamCallback = FunPtr (Ptr () -> CInt -> Ptr () -> IO ())

data StreamFlag

Execution stream creation flags

Constructors

Default

NonBlocking

data StreamWriteFlag

Constructors

WriteValueDefault

WriteValueNoMemoryBarrier

data StreamWaitFlag

Constructors

WaitValueGeq

WaitValueEq

WaitValueAnd

WaitValueNor

WaitValueFlush

data StreamCallbackFlag

create :: [StreamFlag] -> IO Stream

Create a new stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1ga581f0c5833e21ded8b5a56594e243f4

createWithPriority :: StreamPriority -> [StreamFlag] -> IO Stream

Create a stream with the given priority. Work submitted to a higher-priority stream may preempt work already executing in a lower priority stream.

The convention is that lower numbers represent higher priorities. The default priority is zero. The range of meaningful numeric priorities can be queried using getStreamPriorityRange. If the specified priority is outside the supported numerical range, it will automatically be clamped to the highest or lowest number in the range

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g95c1a8c7c3dacb13091692dd9c7f7471

destroy :: Stream -> IO ()

Destroy a stream. If the device is still doing work in the stream when destroy is called, the function returns immediately and the resources associated with the stream will be released automatically once the device has completed all work.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g244c8833de4596bcd31a06cdf21ee758

finished :: Stream -> IO Bool

Check if all operations in the stream have completed.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g1b0d24bbe97fa68e4bc511fb6adfeb0b

block :: Stream -> IO ()

Wait until the device has completed all operations in the Stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g15e49dd91ec15991eb7c0a741beb7dad

callback :: Stream -> StreamCallback -> Ptr () -> [StreamCallbackFlag] -> IO ()

Add a callback to a compute stream. This function will be executed on the host after all currently queued items in the stream have completed.

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g613d97a277d7640f4cb1c03bd51c2483

Since: 0.10.0.0

getFlags :: Stream -> IO [StreamFlag]

Query the flags of a given stream

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g4d39786855a6bed01215c1907fbbfbb7

Since: 0.10.0.0

getPriority :: Stream -> IO StreamPriority

Query the priority of a stream.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g5bd5cb26915a2ecf1921807339488484

getContext :: Stream -> IO Context

Query the context associated with a stream

Requires CUDA-9.2.

https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g5bd5cb26915a2ecf1921807339488484

Since: 0.10.0.0

write :: Storable a => DevicePtr a -> a -> Stream -> [StreamWriteFlag] -> IO ()

Write a value to memory, (presumably) after all preceding work in the stream has completed. Unless the option WriteValueNoMemoryBarrier is supplied, the write is preceded by a system-wide memory fence.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g091455366d56dc2f1f69726aafa369b0

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1gc8af1e8b96d7561840affd5217dd6830

Requires CUDA-8.0 for 32-bit values.

Requires CUDA-9.0 for 64-bit values.

wait :: Storable a => DevicePtr a -> a -> Stream -> [StreamWaitFlag] -> IO ()

Wait on a memory location. Work ordered after the operation will block until the given condition on the memory is satisfied.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g629856339de7bc6606047385addbb398

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g6910c1258c5f15aa5d699f0fd60d6933

Requires CUDA-8.0 for 32-bit values.

Requires CUDA-9.0 for 64-bit values.

defaultStream :: Stream

The default execution stream. This can be configured to have either defaultStreamLegacy or defaultStreamPerThread synchronisation behaviour.

https://docs.nvidia.com/cuda/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

defaultStreamLegacy :: Stream

The legacy default stream is an implicit stream which synchronises with all other streams in the same Context, except for non-blocking streams.

https://docs.nvidia.com/cuda/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

Since: 0.10.0.0

defaultStreamPerThread :: Stream

The per-thread default stream is an implicit stream local to both the thread and the calling Context, and which does not synchronise with other streams (just like explicitly created streams). The per-thread default stream is not a non-blocking stream and will synchronise with the legacy default stream if both are used in the same program.

file:///Developer/NVIDIA/CUDA-9.2/doc/html/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

Since: 0.10.0.0