Foreign.CUDA.Path

Description

Documentation

cudaInstallPath :: FilePath

The base path to the CUDA toolkit installation that this package was compiled against.

cudaBinPath :: FilePath

The path where the CUDA toolkit executables, such as nvcc and ptxas, can be found.

cudaLibraryPath :: FilePath

The path where the CUDA libraries this package was linked against are located

cudaIncludePath :: FilePath

The path where the CUDA headers this package was built against are located