Foreign.CUDA.Runtime.Error

Description

Error handling functions

Documentation

data Status

Return codes from API functions

Constructors

Success

MissingConfiguration

MemoryAllocation

InitializationError

LaunchFailure

PriorLaunchFailure

LaunchTimeout

LaunchOutOfResources

InvalidDeviceFunction

InvalidConfiguration

InvalidDevice

InvalidValue

InvalidPitchValue

InvalidSymbol

MapBufferObjectFailed

UnmapBufferObjectFailed

InvalidHostPointer

InvalidDevicePointer

InvalidTexture

InvalidTextureBinding

InvalidChannelDescriptor

InvalidMemcpyDirection

AddressOfConstant

TextureFetchFailed

TextureNotBound

SynchronizationError

InvalidFilterSetting

InvalidNormSetting

MixedDeviceExecution

CudartUnloading

Unknown

NotYetImplemented

MemoryValueTooLarge

InvalidResourceHandle

NotReady

InsufficientDriver

SetOnActiveProcess

InvalidSurface

NoDevice

ECCUncorrectable

SharedObjectSymbolNotFound

SharedObjectInitFailed

UnsupportedLimit

DuplicateVariableName

DuplicateTextureName

DuplicateSurfaceName

DevicesUnavailable

InvalidKernelImage

NoKernelImageForDevice

IncompatibleDriverContext

PeerAccessAlreadyEnabled

PeerAccessNotEnabled

DeviceAlreadyInUse

ProfilerDisabled

ProfilerNotInitialized

ProfilerAlreadyStarted

ProfilerAlreadyStopped

Assert

TooManyPeers

HostMemoryAlreadyRegistered

HostMemoryNotRegistered

OperatingSystem

PeerAccessUnsupported

LaunchMaxDepthExceeded

LaunchFileScopedTex

LaunchFileScopedSurf

SyncDepthExceeded

LaunchPendingCountExceeded

NotPermitted

NotSupported

HardwareStackError

IllegalInstruction

MisalignedAddress

InvalidAddressSpace

InvalidPc

IllegalAddress

InvalidPtx

InvalidGraphicsContext

NvlinkUncorrectable

StartupFailure

ApiFailureBase

data CUDAException

Constructors

ExitCode Status

UserError String

cudaError :: String -> IO a

Raise a CUDAException in the IO Monad

describe :: Describe a => a -> String

requireSDK :: Name -> Double -> IO a

A specially formatted error message

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise return the error string associated with the return code

nothingIfOk :: Status -> IO ()

Return the error string associated with an unsuccessful return code, otherwise Nothing