Foreign.CUDA.Runtime.Event

Description

Event management for C-for-CUDA runtime environment

Event Management

data Event

Events are markers that can be inserted into the CUDA execution stream and later queried.

data EventFlag

Event creation flags

Constructors

Default

BlockingSync

DisableTiming

data WaitFlag

create :: [EventFlag] -> IO Event

Create a new event

destroy :: Event -> IO ()

Destroy an event

elapsedTime :: Event -> Event -> IO Float

Determine the elapsed time (in milliseconds) between two events

query :: Event -> IO Bool

Determines if a event has actually been recorded

record :: Event -> Maybe Stream -> IO ()

Record an event once all operations in the current context (or optionally specified stream) have completed. This operation is asynchronous.

wait :: Event -> Maybe Stream -> [WaitFlag] -> IO ()

Makes all future work submitted to the (optional) stream wait until the given event reports completion before beginning execution. Synchronisation is performed on the device, including when the event and stream are from different device contexts. Requires cuda-3.2.

block :: Event -> IO ()

Wait until the event has been recorded