Foreign.CUDA.Runtime.Stream

Description

Stream management routines

Stream Management

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

create :: IO Stream

Create a new asynchronous stream

destroy :: Stream -> IO ()

Destroy and clean up an asynchronous stream

finished :: Stream -> IO Bool

Determine if all operations in a stream have completed

block :: Stream -> IO ()

Block until all operations in a Stream have been completed

defaultStream :: Stream

The default execution stream. This can be configured to have either defaultStreamLegacy or defaultStreamPerThread synchronisation behaviour.

https://docs.nvidia.com/cuda/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

defaultStreamLegacy :: Stream

The legacy default stream is an implicit stream which synchronises with all other streams in the same Context, except for non-blocking streams.

https://docs.nvidia.com/cuda/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

Since: 0.10.0.0

defaultStreamPerThread :: Stream

The per-thread default stream is an implicit stream local to both the thread and the calling Context, and which does not synchronise with other streams (just like explicitly created streams). The per-thread default stream is not a non-blocking stream and will synchronise with the legacy default stream if both are used in the same program.

file:///Developer/NVIDIA/CUDA-9.2/doc/html/cuda-driver-api/stream-sync-behavior.html#stream-sync-behavior__default-stream

Since: 0.10.0.0