Foreign.CUDA.Types

Description

Data types that are equivalent and can be shared freely between the CUDA Runtime and Driver APIs.

Pointers

newtype DevicePtr a

A reference to data stored on the device.

Constructors

DevicePtr

Fields

useDevicePtr :: Ptr a

newtype HostPtr a

A reference to page-locked host memory.

A HostPtr is just a plain Ptr, but the memory has been allocated by CUDA into page locked memory. This means that the data can be copied to the GPU via DMA (direct memory access). Note that the use of the system function mlock is not sufficient here --- the CUDA version ensures that the physical address stays this same, not just the virtual address.

To copy data into a HostPtr array, you may use for example withHostPtr together with copyArray or moveArray.

Constructors

HostPtr

Fields

useHostPtr :: Ptr a

Events

newtype Event

Events are markers that can be inserted into the CUDA execution stream and later queried.

Constructors

Event

Fields

useEvent :: Ptr ()

data EventFlag

Event creation flags

Constructors

Default

BlockingSync

DisableTiming

Interprocess

data WaitFlag

Streams

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

type StreamPriority = Int

Priority of an execution stream. Work submitted to a higher priority stream may preempt execution of work already executing in a lower priority stream. Lower numbers represent higher priorities.

data StreamFlag

Possible option flags for stream initialisation. Dummy instance until the API exports actual option values.

defaultStream :: Stream

The main execution stream. No operations overlap with operations in the default stream.