Foreign.CUDA.Driver.Error

Description

Error handling

CUDA Errors

data Status

Constructors

Success

InvalidValue

OutOfMemory

NotInitialized

Deinitialized

ProfilerDisabled

ProfilerNotInitialized

ProfilerAlreadyStarted

ProfilerAlreadyStopped

NoDevice

InvalidDevice

InvalidImage

InvalidContext

ContextAlreadyCurrent

MapFailed

UnmapFailed

ArrayIsMapped

AlreadyMapped

NoBinaryForGPU

AlreadyAcquired

NotMapped

NotMappedAsArray

NotMappedAsPointer

EccUncorrectable

UnsupportedLimit

ContextAlreadyInUse

InvalidSource

FileNotFound

SharedObjectSymbolNotFound

SharedObjectInitFailed

OperatingSystem

InvalidHandle

NotFound

NotReady

LaunchFailed

LaunchOutOfResources

LaunchTimeout

LaunchIncompatibleTexturing

PeerAccessAlreadyEnabled

PeerAccessNotEnabled

PrimaryContextActive

ContextIsDestroyed

Unknown

data CUDAException

Constructors

ExitCode Status

UserError String

describe :: Status -> String

Return a descriptive error string associated with a particular error code

cudaError :: String -> a

Raise a CUDAException. Exceptions can be thrown from pure code, but can only be caught in the IO monad.

cudaErrorIO :: String -> IO a

Raise a CUDAException in the IO Monad

requireSDK :: Name -> Double -> a

A specially formatted error message

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code

nothingIfOk :: Status -> IO ()

Throw an exception with an error string associated with an unsuccessful return code, otherwise return unit.