Foreign.CUDA.Driver.IPC.Event

Description

IPC event management for low-level driver interface.

Restricted to devices which support unified addressing on Linux operating systems.

Since CUDA-4.1.

Documentation

data IPCEvent

A CUDA inter-process event handle.

export :: Event -> IO IPCEvent

Create an inter-process event handle for a previously allocated event. The event must be created with the Interprocess and DisableTiming event flags. The returned handle may then be sent to another process and opened to allow efficient hardware synchronisation between GPU work in other processes.

After the event has been opened in the importing process, record, block, wait, query may be used in either process.

Performing operations on the imported event after the event has been destroyed in the exporting process is undefined.

Requires CUDA-4.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MEM.html#group__CUDA__MEM_1gea02eadd12483de5305878b13288a86c

open :: IPCEvent -> IO Event

Open an inter-process event handle for use in the current process, returning an event that can be used in the current process and behaving as a locally created event with the DisableTiming flag specified.

The event must be freed with destroy. Performing operations on the imported event after the exported event has been destroyed in the exporting process is undefined.

Requires CUDA-4.0.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__MEM.html#group__CUDA__MEM_1gf1d525918b6c643b99ca8c8e42e36c2e