Foreign.CUDA.Driver.Stream

Description

Stream management for low-level driver interface

Stream Management

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

data StreamFlag

Execution stream creation flags

Constructors

StreamDefault

NonBlocking

data StreamWriteFlag

Constructors

WriteValueDefault

WriteValueNoMemoryBarrier

data StreamWaitFlag

Constructors

WaitValueGeq

WaitValueEq

WaitValueAnd

WaitValueFlush

create :: [StreamFlag] -> IO Stream

Create a new stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1ga581f0c5833e21ded8b5a56594e243f4

createWithPriority :: StreamPriority -> [StreamFlag] -> IO Stream

Create a stream with the given priority. Work submitted to a higher-priority stream may preempt work already executing in a lower priority stream.

The convention is that lower numbers represent higher priorities. The default priority is zero. The range of meaningful numeric priorities can be queried using getStreamPriorityRange. If the specified priority is outside the supported numerical range, it will automatically be clamped to the highest or lowest number in the range

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g95c1a8c7c3dacb13091692dd9c7f7471

destroy :: Stream -> IO ()

Destroy a stream. If the device is still doing work in the stream when destroy is called, the function returns immediately and the resources associated with the stream will be released automatically once the device has completed all work.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g244c8833de4596bcd31a06cdf21ee758

finished :: Stream -> IO Bool

Check if all operations in the stream have completed.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g1b0d24bbe97fa68e4bc511fb6adfeb0b

block :: Stream -> IO ()

Wait until the device has completed all operations in the Stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g15e49dd91ec15991eb7c0a741beb7dad

getPriority :: Stream -> IO StreamPriority

Query the priority of a stream.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g5bd5cb26915a2ecf1921807339488484

write :: Storable a => DevicePtr a -> a -> Stream -> [StreamWriteFlag] -> IO ()

Write a value to memory, (presumably) after all preceding work in the stream has completed. Unless the option WriteValueNoMemoryBarrier is supplied, the write is preceded by a system-wide memory fence.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g091455366d56dc2f1f69726aafa369b0

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1gc8af1e8b96d7561840affd5217dd6830

Requires CUDA-8.0 for 32-bit values.

Requires CUDA-9.0 for 64-bit values.

wait :: Storable a => DevicePtr a -> a -> Stream -> [StreamWaitFlag] -> IO ()

Wait on a memory location. Work ordered after the operation will block until the given condition on the memory is satisfied.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g629856339de7bc6606047385addbb398

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g6910c1258c5f15aa5d699f0fd60d6933

Requires CUDA-8.0 for 32-bit values.

Requires CUDA-9.0 for 64-bit values.

defaultStream :: Stream

The main execution stream. No operations overlap with operations in the default stream.