Foreign.CUDA.Driver.Stream

Description

Stream management for low-level driver interface

Stream Management

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

data StreamFlag

Possible option flags for stream initialisation. Dummy instance until the API exports actual option values.

create :: [StreamFlag] -> IO Stream

Create a new stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1ga581f0c5833e21ded8b5a56594e243f4

createWithPriority :: StreamPriority -> [StreamFlag] -> IO Stream

Create a stream with the given priority. Work submitted to a higher-priority stream may preempt work already executing in a lower priority stream.

The convention is that lower numbers represent higher priorities. The default priority is zero. The range of meaningful numeric priorities can be queried using getStreamPriorityRange. If the specified priority is outside the supported numerical range, it will automatically be clamped to the highest or lowest number in the range.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g95c1a8c7c3dacb13091692dd9c7f7471

destroy :: Stream -> IO ()

Destroy a stream. If the device is still doing work in the stream when destroy is called, the function returns immediately and the resources associated with the stream will be released automatically once the device has completed all work.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g244c8833de4596bcd31a06cdf21ee758

finished :: Stream -> IO Bool

Check if all operations in the stream have completed.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g1b0d24bbe97fa68e4bc511fb6adfeb0b

block :: Stream -> IO ()

Wait until the device has completed all operations in the Stream.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g15e49dd91ec15991eb7c0a741beb7dad

getPriority :: Stream -> IO StreamPriority

Query the priority of a stream.

Requires CUDA-5.5.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g5bd5cb26915a2ecf1921807339488484

defaultStream :: Stream

The main execution stream. No operations overlap with operations in the default stream.