Foreign.CUDA.Driver.Texture

Description

Texture management for low-level driver interface

Texture Reference Management

newtype Texture

A texture reference

Constructors

Texture

Fields

useTexture :: Ptr ()

data Format

Texture data formats

Constructors

Word8

Word16

Word32

Int8

Int16

Int32

Half

Float

data AddressMode

Texture reference addressing modes

Constructors

Wrap

Clamp

Mirror

Border

data FilterMode

Texture reference filtering mode

Constructors

Point

Linear

data ReadMode

Texture read mode options

Constructors

ReadAsInteger

NormalizedCoordinates

SRGB

bind :: Texture -> DevicePtr a -> Int64 -> IO ()

Bind a linear array address of the given size (bytes) as a texture reference. Any previously bound references are unbound.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g44ef7e5055192d52b3d43456602b50a8

bind2D :: Texture -> Format -> Int -> DevicePtr a -> (Int, Int) -> Int64 -> IO ()

Bind a linear address range to the given texture reference as a two-dimensional arena. Any previously bound reference is unbound. Note that calls to setFormat can not follow a call to bind2D for the same texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g26f709bbe10516681913d1ffe8756ee2

getAddressMode :: Texture -> Int -> IO AddressMode

Get the addressing mode used by a texture reference, corresponding to the given dimension (currently the only supported dimension values are 0 or 1).

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1gfb367d93dc1d20aab0cf8ce70d543b33

getFilterMode :: Texture -> IO FilterMode

Get the filtering mode used by a texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g2439e069746f69b940f2f4dbc78cdf87

getFormat :: Texture -> IO (Format, Int)

Get the data format and number of channel components of the bound texture.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g90936eb6c7c4434a609e1160c278ae53

setAddressMode :: Texture -> Int -> AddressMode -> IO ()

Specify the addressing mode for the given dimension of a texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g85f4a13eeb94c8072f61091489349bcb

setFilterMode :: Texture -> FilterMode -> IO ()

Specify the filtering mode to be used when reading memory through a texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g595d0af02c55576f8c835e4efd1f39c0

setFormat :: Texture -> Format -> Int -> IO ()

Specify the format of the data and number of packed components per element to be read by the texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g05585ef8ea2fec728a03c6c8f87cf07a

setReadMode :: Texture -> ReadMode -> IO ()

Specify additional characteristics for reading and indexing the texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF.html#group__CUDA__TEXREF_1g554ffd896487533c36810f2e45bb7a28

create :: IO Texture

Deprecated: as of CUDA version 3.2

Create a new texture reference. Once created, the application must call setPtr to associate the reference with allocated memory. Other texture reference functions are used to specify the format and interpretation to be used when the memory is read through this reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF__DEPRECATED.html#group__CUDA__TEXREF__DEPRECATED_1g0084fabe2c6d28ffcf9d9f5c7164f16c

destroy :: Texture -> IO ()

Deprecated: as of CUDA version 3.2

Destroy a texture reference.

http://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__TEXREF__DEPRECATED.html#group__CUDA__TEXREF__DEPRECATED_1gea8edbd6cf9f97e6ab2b41fc6785519d