Foreign.CUDA.Runtime.Device

Description

Device management routines

Device Management

type Device = Int

A device identifier

data DeviceFlag

Device execution flags

Constructors

ScheduleAuto

ScheduleSpin

ScheduleYield

BlockingSync

MapHost

LMemResizeToMax

data DeviceProperties

The properties of a compute device

Constructors

DeviceProperties

Fields

deviceName :: !String

Identifier

computeCapability :: !Compute

Supported compute capability

totalGlobalMem :: !Int64

Available global memory on the device in bytes

totalConstMem :: !Int64

Available constant memory on the device in bytes

sharedMemPerBlock :: !Int64

Available shared memory per block in bytes

regsPerBlock :: !Int

32-bit registers per block

warpSize :: !Int

Warp size in threads (SIMD width)

maxThreadsPerBlock :: !Int

Maximum number of threads per block

maxThreadsPerMultiProcessor :: !Int

Maximum number of threads per multiprocessor

maxBlockSize :: !(Int, Int, Int)

Maximum size of each dimension of a block

maxGridSize :: !(Int, Int, Int)

Maximum size of each dimension of a grid

maxTextureDim1D :: !Int

Maximum texture dimensions

maxTextureDim2D :: !(Int, Int)

maxTextureDim3D :: !(Int, Int, Int)

clockRate :: !Int

Clock frequency in kilohertz

multiProcessorCount :: !Int

Number of multiprocessors on the device

memPitch :: !Int64

Maximum pitch in bytes allowed by memory copies

memBusWidth :: !Int

Global memory bus width in bits

memClockRate :: !Int

Peak memory clock frequency in kilohertz

textureAlignment :: !Int64

Alignment requirement for textures

computeMode :: !ComputeMode

deviceOverlap :: !Bool

Device can concurrently copy memory and execute a kernel

concurrentKernels :: !Bool

Device can possibly execute multiple kernels concurrently

eccEnabled :: !Bool

Device supports and has enabled error correction

asyncEngineCount :: !Int

Number of asynchronous engines

cacheMemL2 :: !Int

Size of the L2 cache in bytes

pciInfo :: !PCI

PCI device information for the device

tccDriverEnabled :: !Bool

Whether this is a Tesla device using the TCC driver

kernelExecTimeoutEnabled :: !Bool

Whether there is a runtime limit on kernels

integrated :: !Bool

As opposed to discrete

canMapHostMemory :: !Bool

Device can use pinned memory

unifiedAddressing :: !Bool

Device shares a unified address space with the host

data Compute

Constructors

Compute !Int !Int

data ComputeMode

The compute mode the device is currently in

Constructors

Default

Exclusive

Prohibited

ExclusiveProcess

choose :: DeviceProperties -> IO Device

Select the compute device which best matches the given criteria

get :: IO Device

Returns which device is currently being used

count :: IO Int

Returns the number of devices available for execution, with compute capability >= 1.0

props :: Device -> IO DeviceProperties

Return information about the selected compute device

set :: Device -> IO ()

Set device to be used for GPU execution

setFlags :: [DeviceFlag] -> IO ()

Set flags to be used for device executions

setOrder :: [Device] -> IO ()

Set list of devices for CUDA execution in priority order

reset :: IO ()

Explicitly destroys and cleans up all runtime resources associated with the current device in the current process. Any subsequent API call will reinitialise the device.

Note that this function will reset the device immediately. It is the caller’s responsibility to ensure that the device is not being accessed by any other host threads from the process when this function is called.

sync :: IO ()

Block until the device has completed all preceding requested tasks. Returns an error if one of the tasks fails.

Peer Access

data PeerFlag

Possible option values for direct peer memory access

accessible :: Device -> Device -> IO Bool

Queries if the first device can directly access the memory of the second. If direct access is possible, it can then be enabled with add. Requires cuda-4.0.

add :: Device -> [PeerFlag] -> IO ()

If the devices of both the current and supplied contexts support unified addressing, then enable allocations in the supplied context to be accessible by the current context. Requires cuda-4.0.

remove :: Device -> IO ()

Disable direct memory access from the current context to the supplied context. Requires cuda-4.0.

Cache Configuration

data Limit

Device limit flags

Constructors

Stacksize

Printffifosize

Mallocheapsize

getLimit :: Limit -> IO Int

Query compute 2.0 call stack limits. Requires cuda-3.1.

setLimit :: Limit -> Int -> IO ()

Set compute 2.0 call stack limits. Requires cuda-3.1.