Foreign.CUDA.Runtime.Stream

Description

Stream management routines

Stream Management

newtype Stream

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

Constructors

Stream

Fields

useStream :: Ptr ()

create :: IO Stream

Create a new asynchronous stream

destroy :: Stream -> IO ()

Destroy and clean up an asynchronous stream

finished :: Stream -> IO Bool

Determine if all operations in a stream have completed

block :: Stream -> IO ()

Block until all operations in a Stream have been completed

defaultStream :: Stream

The main execution stream. No operations overlap with operations in the default stream.