Foreign.CUDA.Runtime.Texture

Description

Texture references

Texture Reference Management

data Texture

Constructors

Texture

Fields

normalised :: !Bool

access texture using normalised coordinates [0.0,1.0)

filtering :: !FilterMode

addressing :: !(AddressMode, AddressMode, AddressMode)

format :: !FormatDesc

data FormatKind

Texture channel format kind

Constructors

Signed

Unsigned

Float

None

data AddressMode

Constructors

Wrap

Clamp

Mirror

Border

data FilterMode

Constructors

Point

Linear

data FormatDesc

A description of how memory read through the texture cache should be interpreted, including the kind of data and the number of bits of each component (x,y,z and w, respectively).

Constructors

FormatDesc

Fields

depth :: !(Int, Int, Int, Int)

kind :: !FormatKind

bind :: String -> Texture -> DevicePtr a -> Int64 -> IO ()

Bind the memory area associated with the device pointer to a texture reference given by the named symbol. Any previously bound references are unbound.

bind2D :: String -> Texture -> DevicePtr a -> (Int, Int) -> Int64 -> IO ()

Bind the two-dimensional memory area to the texture reference associated with the given symbol. The size of the area is constrained by (width,height) in texel units, and the row pitch in bytes. Any previously bound references are unbound.